+++ CHANGELOG.rst
@@ -42,6 +42,8 @@ Breaking changes
 - In ``Group.command`` and ``Group.group`` add type hints and make all arguments
   except ``name`` keyword-only.
 
+- The ``name`` parameter/attribute of ``OptionGroup`` was renamed to ``title``.
+
 Deprecated
 ----------
 Nothing here